Move abstract to outside of directory
